Fix typo. Update license

+++ src/ecdsa.cpp
@@ -1,6 +1,25 @@
+// The MIT License (MIT)
 //
-// Created by Artur Troian on 1/28/17.
+// Copyright (c) 2016 Artur Troian
 //
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
 #include <josepp/crypto.hpp>
 #include <josepp/b64.hpp>
 #include <openssl/err.h>
@@ -29,7 +48,7 @@ std::string ecdsa::sign(const std::string &data)
 
 	uint32_t sig_len;
 
-	if (ECDSA_sign(0, d.data(), d.size(), sig.get(), &sig_len, e_.get()) != 1) {
+	if (ECDSA_sign(0, d.data(), (int)d.size(), sig.get(), &sig_len, e_.get()) != 1) {
 		throw std::runtime_error("Couldn't sign ECDSA");
 	}
 
@@ -42,7 +61,7 @@ bool ecdsa::verify(const std::string &data, const std::string &sig)
 
 	std::vector<uint8_t> s = b64::decode_uri(sig.data(), sig.length());
 
-	if (ECDSA_verify(0, d.data(), d.size(), (const uint8_t *)s.data(), s.size(), e_.get()) != 1) {
+	if (ECDSA_verify(0, d.data(), (int)d.size(), (const uint8_t *)s.data(), (int)s.size(), e_.get()) != 1) {
 		ERR_print_errors_fp(stdout);
 		return false;
 	}
