+++ src/hmac.cpp
@@ -1,6 +1,24 @@
+// The MIT License (MIT)
 //
-// Created by Artur Troian on 1/21/17.
+// Copyright (c) 2016 Artur Troian
 //
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
 
 #include <cstring>
 
@@ -51,7 +69,7 @@ std::string hmac::sign(const std::string &data)
 
 	HMAC_CTX hmac;
 	HMAC_CTX_init(&hmac);
-	HMAC_Init_ex(&hmac, secret_.data(), secret_.length(), evp, NULL);
+	HMAC_Init_ex(&hmac, secret_.data(), (int)secret_.length(), evp, NULL);
 	HMAC_Update(&hmac, (const uint8_t *)data.c_str(), data.size());
 
 	std::shared_ptr<uint8_t> res = std::shared_ptr<uint8_t>(new uint8_t[EVP_MD_size(evp)], std::default_delete<uint8_t[]>());
